Move into Shared SqlDependencyListener.cs (#1308)
